requirements.txt
allure-combine==1.0.11
allure-pytest==2.13.4
base58==2.1.1
black==24.3.0
botocore==1.34.72
boto3==1.34.72
docker==7.0.0
isort==5.13.2
jinja2==3.1.3
neo-mamba==2.6.0
paramiko==3.4.0
pexpect==4.9.0
pytest==7.3.1
PyYAML==6.0.1
python-dotenv==1.0.1
python-dateutil==2.9.0
pytest-lazy-fixture==0.6.3
requests==2.31.0
tenacity==8.2.3
urllib3==2.2.1
